src/fashion_mnist_prediction_service/requirements.txt
flask==2.0.1
werkzeug==2.0.1
pillow==9.0.0
numpy==1.21.0
google-cloud-aiplatform==1.16.0
gunicorn==20.1.0